+++ matchbox-server/src/main/java/ch/ahdis/fhir/hapi/jpa/validation/ImplementationGuideProviderR4.java
@@ -365,7 +365,7 @@ public class ImplementationGuideProviderR4 extends ImplementationGuideResourcePr
 				ig.setDate(npmPackage.getUpdatedTime());
 				ig.setPackageId(npmPackage.getPackageId());
 				if (npmPackage.isCurrentVersion()) {
-					ig.setVersion(npmPackage.getVersionId() + " (current)");
+					ig.setVersion(npmPackage.getVersionId() + " (last)");
 				} else {
 					ig.setVersion(npmPackage.getVersionId());
 				}